requirements.txt
twitchAPI
pyyaml
rich